+++ HISTORY.rst
@@ -10,6 +10,7 @@ History
 * Change Client API
 * Add PreProcess And PostProcess
 * Import Config Object
+* Add Event Listener
 
 0.1.5 (2018-12-14)
 ------------------